+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oauth2/OAuth2Client.kt
@@ -4,7 +4,13 @@ import at.asitplus.openid.*
 import at.asitplus.openid.OpenIdConstants.CODE_CHALLENGE_METHOD_SHA256
 import at.asitplus.openid.OpenIdConstants.GRANT_TYPE_CODE
 import at.asitplus.signum.indispensable.io.Base64UrlStrict
+import at.asitplus.signum.indispensable.josef.JwsHeader
+import at.asitplus.wallet.lib.agent.DefaultCryptoService
+import at.asitplus.wallet.lib.agent.EphemeralKeyWithSelfSignedCert
 import at.asitplus.wallet.lib.iso.sha256
+import at.asitplus.wallet.lib.jws.DefaultJwsService
+import at.asitplus.wallet.lib.jws.JwsContentTypeConstants
+import at.asitplus.wallet.lib.jws.JwsService
 import at.asitplus.wallet.lib.oidvci.DefaultMapStore
 import at.asitplus.wallet.lib.oidvci.MapStore
 import at.asitplus.wallet.lib.oidvci.WalletService
@@ -32,21 +38,39 @@ class OAuth2Client(
      * and then [TokenRequestParameters.codeVerifier], see [RFC 7636](https://datatracker.ietf.org/doc/html/rfc7636).
      */
     private val stateToCodeStore: MapStore<String, String> = DefaultMapStore(),
+    /**
+     * Set this variable to use JAR (JWT-secured authorization requests, RFC 9101)
+     * for PAR (Pushed authorization requests, RFC 9126), as mandated by OpenID4VC HAIP.
+     */
+    val jwsService: JwsService? = DefaultJwsService(DefaultCryptoService(EphemeralKeyWithSelfSignedCert())),
 ) {
 
     /**
-     * Send the result as parameters (either POST or GET) to the server at `/authorize` (or more specific
-     * [OAuth2AuthorizationServerMetadata.authorizationEndpoint]).
+     * Send the result as parameters to the server at [OAuth2AuthorizationServerMetadata.authorizationEndpoint].
+     * Use POST if [OAuth2AuthorizationServerMetadata.pushedAuthorizationRequestEndpoint] is available.
      *
-     * Sample ktor code:
+     * Wraps the actual authorization request in a pushed authorization request (i.e. the `request` property),
+     * if the [jwsService] is available.
+     *
+     * Sample ktor code for GET:
      * ```
      * val authnRequest = client.createAuthRequest(...)
-     * val authnResponse = httpClient.get(issuerMetadata.authorizationEndpointUrl!!) {
+     * httpClient.get(issuerMetadata.authorizationEndpointUrl!!) {
      *     url {
      *         authnRequest.encodeToParameters().forEach { parameters.append(it.key, it.value) }
      *     }
      * }
-     * val authn = AuthenticationResponseParameters.deserialize(authnResponse.bodyAsText()).getOrThrow()
+     * ```
+     *
+     * Sample ktor code for POST:
+     * ```
+     * val authnRequest = client.createAuthRequest(...)
+     * httpClient.submitForm(
+     *     url = issuerMetadata.pushedAuthorizationRequestEndpoint,
+     *     formParameters = parameters {
+     *         authnRequest.encodeToParameters().forEach { append(it.key, it.value) }
+     *     }
+     * )
      * ```
      *
      * @param state to keep internal state in further requests
@@ -55,6 +79,7 @@ class OAuth2Client(
      * @param resource from RFC 8707 Resource Indicators for OAuth 2.0, in OID4VCI flows the value
      * of [IssuerMetadata.credentialIssuer]
      * @param issuerState for OID4VCI flows the value from [CredentialOfferGrantsAuthCode.issuerState]
+     * @param audience for PAR the value of the `issuer` of the Authorization Server
      */
     suspend fun createAuthRequest(
         state: String,
@@ -62,6 +87,7 @@ class OAuth2Client(
         scope: String? = null,
         resource: String? = null,
         issuerState: String? = null,
+        audience: String? = null,
     ) = AuthenticationRequestParameters(
         responseType = GRANT_TYPE_CODE,
         state = state,
@@ -73,11 +99,32 @@ class OAuth2Client(
         redirectUrl = redirectUrl,
         codeChallenge = generateCodeVerifier(state),
         codeChallengeMethod = CODE_CHALLENGE_METHOD_SHA256
+    ).wrapIfNecessary(audience)
+
+    private suspend fun AuthenticationRequestParameters.wrapIfNecessary(audience: String?) =
+        if (jwsService != null) wrapInPar(jwsService, audience) else this
+
+    private suspend fun AuthenticationRequestParameters.wrapInPar(
+        jwsService: JwsService,
+        audience: String?,
+    ) = AuthenticationRequestParameters(
+        clientId = clientId,
+        request = jwsService.createSignedJwsAddingParams(
+            header = JwsHeader(
+                algorithm = jwsService.algorithm,
+                type = JwsContentTypeConstants.OAUTH_AUTHZ_REQUEST
+            ),
+            payload = this.copy(
+                audience = audience,
+                issuer = this.clientId,
+            ),
+            serializer = AuthenticationRequestParameters.serializer(),
+            addJsonWebKey = true,
+        ).getOrThrow().serialize()
     )
 
     /**
-     * Send the result as parameters (either POST or GET) to the server at `/authorize` (or more specific
-     * [OAuth2AuthorizationServerMetadata.authorizationEndpoint]).
+     * Send the result as parameters to the server at [OAuth2AuthorizationServerMetadata.authorizationEndpoint].
      * Use this method if the previous authn request was sent as a pushed authorization request (RFC 9126),
      * and the server has answered with [PushedAuthenticationResponseParameters].
      *
@@ -110,15 +157,15 @@ class OAuth2Client(
          */
         data class PreAuthCode(
             val preAuthorizedCode: String,
-            val transactionCode: String? = null
+            val transactionCode: String? = null,
         ) : AuthorizationForToken()
     }
 
     /**
      * Request token with an authorization code, e.g. from [createAuthRequest], or pre-auth code.
      *
-     * Send the result as POST parameters (form-encoded) to the server at `/token` (or more specific
-     * [OAuth2AuthorizationServerMetadata.tokenEndpoint]).
+     * Send the result as POST parameters (form-encoded) to the server at
+     * [OAuth2AuthorizationServerMetadata.tokenEndpoint].
      *
      * Sample ktor code for authorization code:
      * ```